# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "OKTIOAPI",
    "description": "OKTIO API for reverse auctions",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "x-auth-token",
      "scheme": "bearer",
      "in": "header"
    }
  },
...
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/api/users/getAllUsers": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users in system",
        "responses": {
          "200": {
            "description": "Successful"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/{userId}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get a specific user in the system",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "Get a specific user in the system",
            "schema": {
              "$ref": "#/definitions/userId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/updateProfile"
            }
          }
        }
      }
    },
    "/api/users/updateProfile/{userId}": {
      "patch": {
        "tags": ["Users"],
        "summary": "Update user profile in the system",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "update user profile in the system",
            "schema": {
              "$ref": "#/definitions/updateProfile"
            }
          },
          {
            "name": "updateProfile",
            "in": "body",
            "description": "update user profile in the system",
            "schema": {
              "$ref": "#/definitions/updateProfile"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/updateProfile"
            }
          }
        }
      }
    },
    "/api/users": {
      "post": {
        "tags": ["Users"],
        "summary": "Register a user with a company in the system",
        "parameters": [
          {
            "name": "register",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "$ref": "#/definitions/Register"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/Register"
            }
          }
        }
      }
    },
    "/api/users/changePassword/{userId}": {
      "patch": {
        "tags": ["Users"],
        "summary": "Change a user's password",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "User ID",
            "schema": {
              "$ref": "#/definitions/ChangePassword"
            }
          },
          {
            "name": "new_paaword",
            "in": "body",
            "description": "New password",
            "schema": {
              "$ref": "#/definitions/ChangePassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        }
      }
    },
    "/api/users/verifyEmail/{confirmationCode}": {
      "get": {
        "tags": ["Users"],
        "summary": "User account activation in the system",
        "parameters": [
          {
            "name": "confirmationCode",
            "in": "path",
            "description": "User account verification",
            "schema": {
              "$ref": "#/definitions/verifyEmail"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/verifyEmail"
            }
          }
        }
      }
    },
    "/api/users/ResendConfirmationCode": {
      "post": {
        "tags": ["Users"],
        "summary": "Resend Account Confirmation Code from the system",
        "parameters": [
          {
            "name": "ResendConfirmationCode",
            "in": "body",
            "description": "Sending Verification code to user's email from the system",
            "schema": {
              "$ref": "#/definitions/ResendConfirmationCode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/ResendConfirmationCode"
            }
          }
        }
      }
    },
    "/api/users/verifyForgetPassword": {
      "post": {
        "tags": ["Users"],
        "summary": "Reset user Password on the system",
        "parameters": [
          {
            "name": "forgetPassword",
            "in": "body",
            "description": "Sending Verification code to user's email for the forget password",
            "schema": {
              "$ref": "#/definitions/verifyForgetPassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/verifyForgetPassword"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Register": {
      "required": [
        "first_name",
        "last_name",
        "role",
        "company_name",
        "country",
        "email",
        "password"
      ],
      "properties": {
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "company_name": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "ChangePassword": {
      "required": ["new_password"],
      "properties": {
        "new_password": {
          "type": "string"
        }
      }
    },
    "userId": {
      "required": ["userId"],
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "updateProfile": {
      "required": ["userId"],
      "properties": {
        "image_upload": {
          "type": "string"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "position": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "business_unit": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "mobile": {
          "type": "string"
        },
        "preferences": {
          "type": "object",
          "properties": {
            "language": {
              "type": "string"
            },
            "timezone": {
              "type": "string"
            },
            "currency": {
              "type": "string"
            }
          }
        }
      }
    },
    "verifyEmail": {
      "required": ["confirmationCode"],
      "properties": {
        "confirmationCode": {
          "type": "string"
        }
      }
    },
    "ResendConfirmationCode": {
      "required": ["email"],
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "verifyForgetPassword": {
      "required": ["email"],
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
...
  }
}